# Optimise `surface::pretty`