Fixes very high heap utilization when rendering `surface` AST to `Doc` for pretty printing.

Problem was in `surface::pretty::sequence`, in particular the use of `FlatAlt`. `FlatAlt` requires allocating two AST nodes, one for the single line situation and one for the multi line situation. When used recursively, this will lead to exponential space and time complexity! Fortunately, we can push `FlatAlt` down into the leaves: it is only necessary when deciding whether to emit a separator (eg `,`) or `nil` at the end of the item in the list.