@Kmeakin 2023-01-29
# Benchmark results

## Before
Memory usage, as measured by `heaptrack ./target/release/fathom data --module formats/opentype.fathom /usr/share/fonts/liberation/LiberationMono-Bold.ttf`: 142.9MB

Runtime, as measured by `hyperfine -w 1 -r 10 "./target/release/fathom data --module formats/opentype.fathom /usr/share/fonts/liberation/LiberationMono-Bold.ttf"`:
```
  Time (mean ± σ):     341.7 ms ±   3.3 ms    [User: 315.4 ms, System: 25.9 ms]
  Range (min … max):   337.2 ms … 347.1 ms    10 runs
```

## After
Memory usage, as measured by `heaptrack ./target/release/fathom data --module formats/opentype.fathom /usr/share/fonts/liberation/LiberationMono-Bold.ttf`: 14.2MB

Runtime, as measured by `hyperfine -w 1 -r 10 "./target/release/fathom data --module formats/opentype.fathom /usr/share/fonts/liberation/LiberationMono-Bold.ttf"`:
```
  Time (mean ± σ):      38.0 ms ±   1.0 ms    [User: 33.9 ms, System: 4.2 ms]
  Range (min … max):    37.1 ms …  39.7 ms    10 runs
```